==> notebooks/03_3_scatter_Viz.ipynb <==
# import itertools
import os
import sys
from pathlib import Path

import cv2
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import skimage.io

from collections import defaultdict
from tqdm.auto import tqdm
from joblib import Parallel, delayed
import re
import h5py
import napari
import tifffile as tiff
import seaborn as sns
import pickle

%load_ext autoreload
%autoreload 2

p_dir = (Path().cwd().parents[0]).absolute()

module_path = str(p_dir / "src")
 
if module_path not in sys.path:
    sys.path.append(module_path)

data_dir = (Path().cwd().parents[0] / 'data').absolute()

df_meta_path = data_dir / 'sensitivity_v2' / 'metadata' / 'imgs_reg.csv'
df_imgs = pd.read_csv(df_meta_path)


df_imgs

# Plot examples

Revision Figure on tratement 
Need to look at FOV 6 and 7 and compared control and treated

import napari 

# remove first and last consecutive zeros along dimension
def removeFirstAndLast(zeroArray):
    
    # remove first
    for ii, val in enumerate(zeroArray):
        if np.sum(zeroArray[:ii + 1]) != ii + 1: # all consecutive trues
            break # exit loop
            
        else:
            lowest = ii # index of last of the first consecutive zeros
            
    # remove last
    for ii, val in enumerate(np.flip(zeroArray)):
        if np.sum(zeroArray[:ii + 1]) != ii + 1: # all consecutive trues
            break # exit loop
            
        else:
            highest = len(zeroArray) - ii # index of first of the last consecutive zeros            
    
    return (lowest, highest)

# remove black borders from napari screenshot (2D image)
def removeBorders(RGB):
    rowMin = []
    rowMax = []
    colMin = []
    colMax = []
    for jj in range(3): # each RGB
        
        img = RGB[:, :, jj]
        # find first and last rows and cols of all zeros
        # remove any rows of zeros
        zeroRows = np.all(img == 0, axis = 1)
        zeroCols = np.all(img == 0, axis = 0)

        posRows = removeFirstAndLast(zeroRows)
        posCols = removeFirstAndLast(zeroCols)            
        
        rowMin.append(posRows[0])
        rowMax.append(posRows[1])
        colMin.append(posCols[0])
        colMax.append(posCols[1])
        
    # make sure border signal is not cut out
    rowMin = np.min(rowMin)
    rowMax = np.max(rowMax)
    colMin = np.min(colMin)
    colMax = np.max(colMax)
        
    cropped = RGB[rowMin: rowMax, colMin: colMax, :] 
    
    return cropped

def get_loc_ppi(df, Condition, FOV, x_min, y_min, x_max, y_max, PPI_name):
    df_ppi = df[(df.Condition == Condition) & (df.FOV == FOV)]
    df_ppi = df_ppi[df_ppi.PPI == PPI_name]
    df_ppi = df_ppi[(df_ppi.column > x_min) & (df_ppi.column < x_max) 
                    & (df_ppi.row > y_min) & (df_ppi.row < y_max)] 
    points = df_ppi[['row', 'column']].to_numpy()
    points = points - np.array([y_min, x_min])
    return points

# Read PPI 
PPI_save_path =  data_dir / 'sensitivity_v2'/ 'PPI'

dfs = []
for path in os.listdir(PPI_save_path):
    if 'csv' in path:
        df = pd.read_csv(PPI_save_path / path)
        dfs.append(df)

df = pd.concat(dfs)
df = df.rename(columns={'x': 'row', 'y': 'column'})
df.head()

image_dir = (Path().cwd().parents[0] / 'figures').absolute() / 'PLA_scatter'
image_dir.mkdir(parents=True, exist_ok=True)

## B05

# Get images
index = 4
row = df_imgs.iloc[index]
print(row)
path = row.Path

with h5py.File(path, 'r') as f:
    imgs = f['imgs'][:]
    markers = f['imgs'].attrs['Marker']


# Define matching bbox to figures
x_min, y_min = 1292, 712
x_max, y_max = 1720, 1204
DAPI_ch = 0

# Get DAPI image
dapiImg = imgs[0, y_min:y_max, x_min:x_max]

# Get ppi location
PPI_names = ['Sox2/Oct4', 'p-P90rsk/NF-Kb', 'Cyclin D1/CDK4', 'Bim/Tom20']
points = [get_loc_ppi(df, row.Condition, row.FOV, x_min, y_min, x_max, y_max, PPI_name) for PPI_name in PPI_names]


# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[0], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[1], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_B05_1.tif', removeBorders(screenshot))

# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[2], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[3], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_B05_2.tif', removeBorders(screenshot))

## C05

# Get images
index = 0
row = df_imgs.iloc[index]
print(row)
path = row.Path

with h5py.File(path, 'r') as f:
    imgs = f['imgs'][:]
    markers = f['imgs'].attrs['Marker']


# Define matching bbox to figures
x_min, y_min = 1464,2408
x_max, y_max = 1852,2824
DAPI_ch = 0

# Get DAPI image
dapiImg = imgs[0, y_min:y_max, x_min:x_max]

# Get ppi location
PPI_names = ['Sox2/Oct4', 'p-P90rsk/NF-Kb', 'Cyclin D1/CDK4', 'Bim/Tom20']
points = [get_loc_ppi(df, row.Condition, row.FOV, x_min, y_min, x_max, y_max, PPI_name) for PPI_name in PPI_names]


# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[0], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[1], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_C05_1.tif', removeBorders(screenshot))

# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[2], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[3], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_C05_2.tif', removeBorders(screenshot))

## D05

# Get images
index = 2
row = df_imgs.iloc[index]
print(row)
path = row.Path

with h5py.File(path, 'r') as f:
    imgs = f['imgs'][:]
    markers = f['imgs'].attrs['Marker']


# Define matching bbox to figures
x_min, y_min = 400, 1580
x_max, y_max = 740, 1892

# Get DAPI image
dapiImg = imgs[0, y_min:y_max, x_min:x_max]

# Get ppi location
PPI_names = ['Sox2/Oct4', 'p-P90rsk/NF-Kb', 'Cyclin D1/CDK4', 'Bim/Tom20']
points = [get_loc_ppi(df, row.Condition, row.FOV, x_min, y_min, x_max, y_max, PPI_name) for PPI_name in PPI_names]


# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[0], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[1], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_D05_1.tif', removeBorders(screenshot))

# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[2], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[3], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_D05_2.tif', removeBorders(screenshot))

## E05

# Get images
index = 1
row = df_imgs.iloc[index]
print(row)
path = row.Path

with h5py.File(path, 'r') as f:
    imgs = f['imgs'][:]
    markers = f['imgs'].attrs['Marker']


# Define matching bbox to figures
x_min, y_min = 2198, 2712
x_max, y_max = 2546, 3012
DAPI_ch = 0

# Get DAPI image
dapiImg = imgs[0, y_min:y_max, x_min:x_max]

# Get ppi location
PPI_names = ['Sox2/Oct4', 'p-P90rsk/NF-Kb', 'Cyclin D1/CDK4', 'Bim/Tom20']
points = [get_loc_ppi(df, row.Condition, row.FOV, x_min, y_min, x_max, y_max, PPI_name) for PPI_name in PPI_names]


# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[0], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[1], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_E05_1.tif', removeBorders(screenshot))

# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[2], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[3], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_E05_2.tif', removeBorders(screenshot))

## F05

# Get images
index = 3
row = df_imgs.iloc[index]
print(row)
path = row.Path

with h5py.File(path, 'r') as f:
    imgs = f['imgs'][:]
    markers = f['imgs'].attrs['Marker']


# Define matching bbox to figures
x_min, y_min = 772, 584
x_max, y_max = 1148, 960
DAPI_ch = 0

# Get DAPI image
dapiImg = imgs[0, y_min:y_max, x_min:x_max]

# Get ppi location
PPI_names = ['Sox2/Oct4', 'p-P90rsk/NF-Kb', 'Cyclin D1/CDK4', 'Bim/Tom20']
points = [get_loc_ppi(df, row.Condition, row.FOV, x_min, y_min, x_max, y_max, PPI_name) for PPI_name in PPI_names]


# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[0], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[1], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_F05_1.tif', removeBorders(screenshot))

# Napari viewer
viewer = napari.view_image(dapiImg, colormap='blue')
points_layer = viewer.add_points(points[2], size=8,face_color='magenta',)
points_layer = viewer.add_points(points[3], size=8,face_color='lime',)

# Get screenshot
screenshot = viewer.screenshot(size=np.array(dapiImg.shape)*2)
viewer.close()
# Save image
tiff.imsave(image_dir / f'sensitivity_F05_2.tif', removeBorders(screenshot))
