==> notebooks/04_CellCulutre_MLP_count_3D.ipynb <==
# import itertools
import os
import sys
from pathlib import Path

import cv2
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import skimage.io

from collections import defaultdict
from tqdm.auto import tqdm
from joblib import Parallel, delayed
import re
import h5py
import napari
import tifffile as tiff
import seaborn as sns
import pickle

import networkx as nx
from sklearn.model_selection import train_test_split

import torch_geometric.utils
import torch

%load_ext autoreload
%autoreload 2

p_dir = (Path().cwd().parents[0]).absolute()

module_path = str(p_dir / "src")

if module_path not in sys.path:
    sys.path.append(module_path)

data_dir = (Path().cwd().parents[0] / 'data').absolute()

import lightning.pytorch as pl
import PPIGraph

# Test custom torch dataset

from torch_geometric.loader import DataLoader


condition_mapping = {'HCC827Ctrl': 0, 'HCC827Osim': 1}


graph_path = data_dir / 'OCT Cell Culture' / '3D_Whole' / 'graphs' 

dataset = PPIGraph.GraphDatasetMLP_sub(graph_path, 'raw', 'pt', condition_mapping=condition_mapping, n_c=2)
train_set, val_set, test_set = PPIGraph.train_test_val_split(dataset)

# Create Dataloader
train_loader = DataLoader(train_set, batch_size=32, shuffle=True)
val_loader = DataLoader(val_set, batch_size=32, shuffle=True)
test_loader = DataLoader(test_set, batch_size=32, shuffle=False)


print(f'Dataset: {dataset}:')
print('======================')
print(f'Number of graphs: {len(dataset)}')
print(f'Number of features: {dataset.num_features}')
print(f'Number of classes: {dataset.num_classes}')

print(f'Train set: {len(train_set)}, test set: {len(test_set)}, val set: {len(val_set)}')

for step, data in enumerate(train_loader):
    print(f'Step {step + 1}:')
    print('=======')
    print(f'Number of graphs in the current batch: {data.num_graphs}')
    print(data)
    print()
    break

# Lightning training

import itertools
from lightning.pytorch.accelerators import find_usable_cuda_devices
import wandb

# condition = 'base'
# checkpoint_folder = (Path().cwd().parents[0]).absolute() / 'data' / "saved_models" / f'MLP_{condition}' 
# project_name = f'PLA_041623_{condition}'

# AVAIL_GPUS = [1]
# BATCH_SIZE = 64 if AVAIL_GPUS else 32

# # Setting the seed
# pl.seed_everything(42)

# num_layers = [2,3,4]
# hiddens = [8, 16, 32, 64]
# pool = 'mean'

# epochs = 30

# for NUM_LAYERS, HIDDEN_CHANNELS in list(itertools.product(*[num_layers, hiddens])):
    
#     # Path to the folder where the pretrained models are saved
#     CHECKPOINT_PATH = checkpoint_folder / f'_{NUM_LAYERS}_{HIDDEN_CHANNELS}_count_cyto'
#     CHECKPOINT_PATH.mkdir(parents=True, exist_ok=True)
    
#     # Run training
#     models = ['MLP']
#     for model_name in models:
#         run = wandb.init(project=project_name, name=model_name+f'_{NUM_LAYERS}_{HIDDEN_CHANNELS}_count_cyto')
#         model, result, trainer = PPIGraph.train_graph_classifier(model_name, train_set, val_set, test_set, 
#                                                                  dataset, CHECKPOINT_PATH, AVAIL_GPUS, 
#                                                                  in_channels=HIDDEN_CHANNELS,
#                                                                  hidden_channels=HIDDEN_CHANNELS, 
#                                                                  out_channels=HIDDEN_CHANNELS, 
#                                                                  num_layers=NUM_LAYERS, 
#                                                                  epochs=epochs,
#                                                                  embedding=True,
#                                                                  graph_pooling=pool)
#         run.finish()

# for NUM_LAYERS, HIDDEN_CHANNELS in list(itertools.product(*[num_layers, hiddens])):
    
#     # Path to the folder where the pretrained models are saved
#     CHECKPOINT_PATH = checkpoint_folder / f'_{NUM_LAYERS}_{HIDDEN_CHANNELS}_count_cell'
#     CHECKPOINT_PATH.mkdir(parents=True, exist_ok=True)
    
#     # Run training
#     models = ['MLP']
#     for model_name in models:
#         run = wandb.init(project=project_name, name=model_name+f'_{NUM_LAYERS}_{HIDDEN_CHANNELS}_count_cell')
#         model, result, trainer = PPIGraph.train_graph_classifier(model_name, train_set, val_set, test_set, 
#                                                                  dataset, CHECKPOINT_PATH, AVAIL_GPUS, 
#                                                                  in_channels=HIDDEN_CHANNELS,
#                                                                  hidden_channels=HIDDEN_CHANNELS, 
#                                                                  out_channels=HIDDEN_CHANNELS, 
#                                                                  num_layers=NUM_LAYERS, 
#                                                                  epochs=epochs,
#                                                                  embedding=True,
#                                                                  graph_pooling=pool)
#         run.finish()

# K Fold

import itertools
from lightning.pytorch.accelerators import find_usable_cuda_devices
import wandb

from sklearn.model_selection import KFold, StratifiedKFold
from torch.utils.data import SubsetRandomSampler

# Filter out by maximum number of counts per cell
min_count = 20
max_count = 70

graph_path = data_dir / 'OCT Cell Culture' / '3D_Whole' / 'graphs' 

dataset_subcellular = PPIGraph.GraphDataset(graph_path, 'raw', 'pt', condition_mapping=condition_mapping, n_c=2)

# Create Dataloader
loader = DataLoader(dataset_subcellular, batch_size=1, shuffle=False)

# Get Indices
indices = []
for step, data in enumerate(loader):
    if len(data.x) <= min_count:
        continue 
    
    if (data.x.sum(axis=0) >= max_count).any():
        continue
    indices.append(step)
    
# Get subset dataset
dataset_filtered = dataset.index_select(indices)


len(dataset_filtered)

condition = 'Kfold'
checkpoint_folder = (Path().cwd().parents[0]).absolute() / 'data' / "saved_models" / f"MLP_count_{condition}" 
project_name = f'PLA_050823_{condition}'

AVAIL_GPUS = [1]

# Setting the seed
pl.seed_everything(42)

NUM_LAYERS = 2
HIDDEN_CHANNELS = 16

epochs = 50

# K fold nuclei cyto seperated
k_folds = 5
kfold = KFold(n_splits=k_folds, shuffle=True, random_state=42)

for fold, (train_ids, valid_ids) in enumerate(kfold.split(dataset_filtered)):
    train_subset = dataset_filtered.index_select(train_ids.tolist())
    val_subset = dataset_filtered.index_select(valid_ids.tolist())
    
    # Path to the folder where the pretrained models are saved
    CHECKPOINT_PATH = checkpoint_folder / f'3D_{NUM_LAYERS}_{HIDDEN_CHANNELS}_cyto_{fold}' 
    CHECKPOINT_PATH.mkdir(parents=True, exist_ok=True)

    # Skip already trained kfold and pool
    checkpoint = CHECKPOINT_PATH / "GraphLevelMLP" / "GraphLevelMLP.ckpt" 
    if checkpoint.exists():
        print(checkpoint)
        continue

    # Run training
    run = wandb.init(project=project_name, name=f'3D_MLP_{NUM_LAYERS}_{HIDDEN_CHANNELS}_cyto_{fold}', 
                    group=f'3D_MLP_cyto')
    PPIGraph.train_graph_classifier_kfold('MLP', 
                                         train_subset, 
                                         val_subset, 
                                         dataset, 
                                         CHECKPOINT_PATH, 
                                         AVAIL_GPUS, 
                                         in_channels=10,
                                         hidden_channels=HIDDEN_CHANNELS, 
                                         out_channels = HIDDEN_CHANNELS,
                                         num_layers=NUM_LAYERS, 
                                         epochs=epochs,
                                         embedding=False,
                                         batch_size=128,
                                         graph_pooling='mean')
    run.finish()
