==> notebooks/18_xenograft_ML_benchmark.ipynb <==
import itertools
import os
import sys
from pathlib import Path

import cv2
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import skimage.io

from collections import defaultdict
from tqdm.auto import tqdm
from joblib import Parallel, delayed
import re
import h5py
import napari
import seaborn as sns
from sklearn.preprocessing import StandardScaler


%load_ext autoreload
%autoreload 2

p_dir = (Path().cwd().parents[0]).absolute()

module_path = str(p_dir / "src")

if module_path not in sys.path:
    sys.path.append(module_path)

data_dir = (Path().cwd().parents[0] / 'data').absolute()

import PPIGraph

# Read per cell data

from torch_geometric.loader import DataLoader

# Define condition mapping
condition_mapping = {'1W': 0, '1M': 1}

# Load graph dataset and process if neede
graph_path = data_dir / 'OCT mouse' / 'graphs' 

# Filter out by maximum number of counts per cell
min_count = 5
max_count = 40

dataset = PPIGraph.GraphDataset(graph_path, 'raw', 'pt',  condition_mapping=condition_mapping, n_c=2)

# Create Dataloader
loader = DataLoader(dataset, batch_size=1, shuffle=False)

# # Get Indices
# indices = []
# for step, data in enumerate(loader):
#     if len(data.x) <= min_count:
#         continue 
    
#     if (data.x.sum(axis=0) >= max_count).any():
#         continue
#     indices.append(step)
    
# # Get subset dataset
# dataset_filtered = dataset.index_select(indices)


# np.save(graph_path / 'indices.npy', indices)
# indices_test = np.load(graph_path / 'indices.npy')

# np.array_equal(indices_test,np.array(indices))

dataset

indices = np.load(graph_path / 'indices.npy')
indices = indices[::15]
dataset_filtered = dataset.index_select(list(indices))


dataset_filtered

dataset_filtered[0]

# ML model

from sklearn.svm import SVC
from sklearn.ensemble import RandomForestClassifier, AdaBoostClassifier, GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn import preprocessing, metrics
from sklearn.neural_network import MLPClassifier
from sklearn.linear_model import LogisticRegression

from sklearn.model_selection import train_test_split, KFold
from sklearn.utils.class_weight import compute_class_weight
import wandb

# condition = 'ML'
# project_name = f'PLA_xeno_11324_{condition}'

# models = {
#     'Adaboost': AdaBoostClassifier(),
#     'DecisionTree': DecisionTreeClassifier(),
#     'GradientBoosting' : GradientBoostingClassifier(),
#     'NaiveBayes': GaussianNB(),
#     'RandomForest': RandomForestClassifier(), 
#     'SVM': SVC(probability =True),
#     'LogisticRegression':  LogisticRegression(),
#     'MLP': MLPClassifier(random_state=1, max_iter=100, hidden_layer_sizes=[16, 16, 16])
# }

# import torch 

# data = torch.empty((0, 5))
# label = []

# for graph in dataset_filtered:
#     new_row =graph.x.sum(axis = 0).unsqueeze(0)
#     data = torch.cat((data, new_row), dim=0)
#     label.append(graph.condition)


# np.save(graph_path / 'data.npy', data)
# np.save(graph_path / 'label.npy', label)

# scaler = StandardScaler()

# # Run model on cell count
# X = data.numpy()
# X = scaler.fit_transform(X)
# y = np.array(label)

# # X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=0)
# kfold = KFold(n_splits = 10, shuffle = True, random_state = 0)

# for model_name, model in models.items():
#     for k, (train_index, test_index) in enumerate(kfold.split(X)):
#         # Split the dataset
#         X_train, X_test = X[train_index], X[test_index]
#         y_train, y_test = y[train_index], y[test_index]
        
#         model.fit(X_train, y_train)
#         y_pred = model.predict(X_test)
#         y_probas = model.predict_proba(X_test)

#         run = wandb.init(project=project_name, group=model_name+'_cell', name=model_name+f'_cell_{k}')

#         accuracy = metrics.accuracy_score(y_test, y_pred)
#         b_accuracy = metrics.balanced_accuracy_score(y_test, y_pred)
#         f1 = metrics.f1_score(y_test, y_pred)
#         auc = metrics.roc_auc_score(y_test, y_pred)
#         wandb.log({"accuracy": accuracy, 'b_accuracy': b_accuracy, 'f1':f1, 'auc': auc})
#     run.finish()

# import torch 

# data = torch.empty((0, 10))
# label = []

# for graph in dataset_filtered:
#     cell =graph.x.sum(axis = 0).unsqueeze(0)

#     nuclei = torch.tile(graph.nuclei, (5, 1)).T * graph.x
#     nuclei = nuclei.sum(axis=0).unsqueeze(0)
#     cyto = cell - nuclei
#     new_row = torch.cat([cyto, nuclei], axis=1)
#     data = torch.cat((data, new_row), dim=0)
#     label.append(graph.condition)

# scaler = StandardScaler()

# # Run model on cell count
# X = data.numpy()
# X = scaler.fit_transform(X)
# y = np.array(label)

# # X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.4, random_state=0)
# kfold = KFold(n_splits = 5, shuffle = True, random_state = 0)

# for model_name, model in models.items():
#     for k, (train_index, test_index) in enumerate(kfold.split(X)):
#         # Split the dataset
#         X_train, X_test = X[train_index], X[test_index]
#         y_train, y_test = y[train_index], y[test_index]
        
#         model.fit(X_train, y_train)
#         y_pred = model.predict(X_test)
#         y_probas = model.predict_proba(X_test)

#         run = wandb.init(project=project_name, group=model_name+'_cytonuclei', name=model_name+f'_cytonuclei_{k}')

#         accuracy = metrics.accuracy_score(y_test, y_pred)
#         b_accuracy = metrics.balanced_accuracy_score(y_test, y_pred)
#         f1 = metrics.f1_score(y_test, y_pred)
#         auc = metrics.roc_auc_score(y_test, y_pred)
#         wandb.log({"accuracy": accuracy, 'b_accuracy': b_accuracy, 'f1':f1, 'auc': auc})
#     run.finish()

# GNN training

import lightning.pytorch as pl
import PPIGraph
from sklearn.model_selection import train_test_split, KFold


import itertools
from lightning.pytorch.accelerators import find_usable_cuda_devices
import wandb

dataset_name = 'xenograph'
condition = 'Kfold'
checkpoint_folder = (Path().cwd().parents[0]).absolute() / 'data' /'OCT mouse' /"saved_models" / dataset_name / f"Graph_GNNs_{condition}" 
project_name = f'PLA_01122024_{dataset_name}_{condition}'

AVAIL_GPUS = [1]

# Setting the seed
pl.seed_everything(42)
epochs = 20

# NUM_LAYERS = 2
# HIDDEN_CHANNELS = 16
# pools = ['mean', 'max', 'sum', 'attention', 'attention2']
# num_layers = [2,3,4]
# hiddens = [16, 32, 64]


# # model = 'GAT'
# model = 'GINConv'
params = [[3,32,'mean','MLP'], 
[3,32,'attention2','GraphConv'],
[2,64,'attention2','GCN'],
[3,64,'attention','GAT'],
[2,16,'attention','GINConv']]

import torch
torch.set_float32_matmul_precision('medium')

k_folds = 5
kfold = KFold(n_splits=k_folds, shuffle=True, random_state=42)

for fold, (train_ids, valid_ids) in enumerate(kfold.split(dataset_filtered)):
    train_subset = dataset_filtered.index_select(train_ids.tolist())
    val_subset = dataset_filtered.index_select(valid_ids.tolist())
    for NUM_LAYERS, HIDDEN_CHANNELS, pool, model in params:
            # Path to the folder where the pretrained models are saved
        # CHECKPOINT_PATH = checkpoint_folder / f'GAT_{NUM_LAYERS}_{HIDDEN_CHANNELS}_onehot_{fold}' / pool
        CHECKPOINT_PATH = checkpoint_folder / f'{model}_{NUM_LAYERS}_{HIDDEN_CHANNELS}_onehot_{fold}' / pool
        CHECKPOINT_PATH.mkdir(parents=True, exist_ok=True)

        # Skip already trained kfold and pool
        checkpoint = CHECKPOINT_PATH / f"GraphLevel{model}" / f"GraphLevel{model}.ckpt" 
        if checkpoint.exists():
            print(checkpoint)
            continue

        # Run training
        run = wandb.init(project=project_name, name=f'{model}_{NUM_LAYERS}_{HIDDEN_CHANNELS}_onehot_{fold}', 
                        group=f'{model}_{pool}', 
                        # mode="disabled"
                        )
        PPIGraph.train_graph_classifier_kfold(model, 
                                                train_subset, 
                                                val_subset, 
                                                dataset_filtered, 
                                                CHECKPOINT_PATH, 
                                                AVAIL_GPUS, 
                                                in_channels=5,
                                                hidden_channels=HIDDEN_CHANNELS, 
                                                out_channels = HIDDEN_CHANNELS,
                                                num_layers=NUM_LAYERS, 
                                                epochs=epochs,
                                                embedding=False,
                                                batch_size=256,
                                                num_workers= 24,
                                                graph_pooling=pool)
        run.finish()
